Refactor provision code into library
